+++ cubrid/CMakeLists.txt
@@ -335,6 +335,7 @@ set(TRANSACTION_SOURCES
   ${TRANSACTION_DIR}/log_recovery_redo.cpp
   ${TRANSACTION_DIR}/log_recovery_redo_parallel.cpp
   ${TRANSACTION_DIR}/log_replication.cpp
+  ${TRANSACTION_DIR}/log_replication_jobs.cpp
   ${TRANSACTION_DIR}/log_replication_atomic.cpp
   ${TRANSACTION_DIR}/log_replication_mvcc.cpp
   ${TRANSACTION_DIR}/log_storage.cpp
@@ -374,6 +375,8 @@ set(TRANSACTION_HEADERS
   ${TRANSACTION_DIR}/log_recovery_redo.hpp
   ${TRANSACTION_DIR}/log_recovery_redo_parallel.hpp
   ${TRANSACTION_DIR}/log_replication.hpp
+  ${TRANSACTION_DIR}/log_replication_jobs.hpp
+  ${TRANSACTION_DIR}/log_replication.cpp.hpp
   ${TRANSACTION_DIR}/log_replication_atomic.hpp
   ${TRANSACTION_DIR}/log_replication_mvcc.hpp
   ${TRANSACTION_DIR}/log_storage.hpp
